Cleanup repo and add env template
