+++ .env.example
@@ -1,0 +1,4 @@
+DATABASE_PATH=cryptopilot.sqlite
+EXPRESS_SERVER_URL=http://localhost:5002
+FLASK_SERVER_URL=http://localhost:5001
+TEST_DATABASE_PATH=cryptopilot_test.sqlite
